# pygmt.filter1d: Improve performance by storing output in virtual files