--- a/pygmt/src/filter1d.py
+++ b/pygmt/src/filter1d.py
@@ -104,6 +104,7 @@ def filter1d(
     -------
     ret
         Return type depends on ``outfile`` and ``output_type``:
+
         - None if ``outfile`` is set (output will be stored in file set by ``outfile``)
         - :class:`pandas.DataFrame` or :class:`numpy.ndarray` if ``outfile`` is not set
           (depends on ``output_type``)