--- a/pygmt/tests/test_filter1d.py
+++ b/pygmt/tests/test_filter1d.py
@@ -2,7 +2,6 @@
 Test pygmt.filter1d.
 """
 
-import numpy as np
 import pandas as pd
 import pytest
 from pygmt import filter1d
@@ -17,25 +16,11 @@ def fixture_data():
     return load_sample_data(name="maunaloa_co2")
 
 
-def test_filter1d_no_outfile(data):
+@pytest.mark.benchmark
+def test_filter1d(data):
     """
-    Test filter1d with no set outfile.
+    Test the basic functionality of filter1d.
     """
     result = filter1d(data=data, filter_type="g5")
+    assert isinstance(result, pd.DataFrame)
     assert result.shape == (671, 2)
-
-
-@pytest.mark.benchmark
-def test_filter1d_format(data):
-    """
-    Test that correct formats are returned.
-    """
-    time_series_default = filter1d(data=data, filter_type="g5")
-    assert isinstance(time_series_default, pd.DataFrame)
-    assert time_series_default.shape == (671, 2)
-    time_series_array = filter1d(data=data, filter_type="g5", output_type="numpy")
-    assert isinstance(time_series_array, np.ndarray)
-    assert time_series_array.shape == (671, 2)
-    time_series_df = filter1d(data=data, filter_type="g5", output_type="pandas")
-    assert isinstance(time_series_df, pd.DataFrame)
-    assert time_series_df.shape == (671, 2)