**Preview**: https://pygmt-dev--3085.org.readthedocs.build/en/3085/api/generated/pygmt.filter1d.html#pygmt.filter1d

Similar to #3097.