@seisman 2024-03-12
Benchmarks:
```
>>> import numpy as np
>>> for n in [10, 100, 1000, 10000, 100000, 1000000]:
...     time = np.arange(0, n, 1)
...     data = np.random.random(n)
...     np.savetxt(fname=f"test-{n}.txt", X=np.array([time, data]).T, fmt="%.2f")
>>> import pygmt
>>> for n in [10, 100, 1000, 10000, 100000, 1000000]:
...     %timeit pygmt.filter1d(f"test-{n}.txt", filter_type="g5")
```

Main branch:
```
4.48 ms ± 16.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.63 ms ± 16.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
6 ms ± 9.02 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
23 ms ± 126 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
168 ms ± 798 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.68 s ± 5.18 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

This branch:
```
4.1 ms ± 14 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.15 ms ± 12.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.59 ms ± 35.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
8.91 ms ± 272 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
54.3 ms ± 118 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
504 ms ± 342 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```